+++ examples/admin-sdk-app/src/frontend/cms/ex-dailymotion/ex-dailymotion-config.vue
@@ -13,7 +13,7 @@
 
 <script setup lang="ts">
 import { onBeforeMount, ref, computed } from 'vue';
-import { data } from '@shopware-ag/meteor-admin-sdk';
+import { data } from '@haokeyingxiao/meteor-admin-sdk';
 import EX_DAILYMOTION_CONSTANTS from './ex-dailymotion-constants';
 
 const dailyUrlValue = ref('');