+++ examples/admin-sdk-app/src/frontend/locations/module/example-module/ex-cms-card.vue
@@ -11,8 +11,8 @@
 </template>
 
 <script setup lang="ts">
-import { window } from '@shopware-ag/meteor-admin-sdk';
-import { SwCard } from '@shopware-ag/meteor-component-library';
+import { window } from '@haokeyingxiao/meteor-admin-sdk';
+import { SwCard } from '@haokeyingxiao/meteor-component-library';
 
 function goToCmsModule() {
   window.routerPush({