+++ examples/admin-sdk-plugin/src/Resources/app/administration/src/views/location/index.ts
@@ -1,6 +1,6 @@
 import { defineComponent } from 'vue'
-import { location } from "@shopware-ag/meteor-admin-sdk";
-import { MtButton, MtNumberField } from '@shopware-ag/meteor-component-library';
+import { location } from "@haokeyingxiao/meteor-admin-sdk";
+import { MtButton, MtNumberField } from '@haokeyingxiao/meteor-component-library';
 
 export default defineComponent({
     components: {