+++ examples/admin-sdk-plugin/src/Resources/app/administration/src/views/ui/main-module/add-main-module.ts
@@ -1,6 +1,6 @@
 import { defineComponent } from 'vue';
-import { notification, context, data, window } from '@shopware-ag/meteor-admin-sdk';
-import { SwButton, SwTextField, SwTextarea } from '@shopware-ag/meteor-component-library';
+import { notification, context, data, window } from '@haokeyingxiao/meteor-admin-sdk';
+import { SwButton, SwTextField, SwTextarea } from '@haokeyingxiao/meteor-component-library';
 
 const { repository, Classes: { Criteria } } = data;